update unlock screen, refactor + added autofocus and some other small things
