+++ lib/view_models/lock_screen_view_model.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 class LockScreenViewModel extends ChangeNotifier {
   String _username = 'User001';
-  String _profileImage = 'assets/images/onboarding_bg_1.svg'; // Placeholder image
+  String _profileImage = 'assets/images/avatar.png'; // Placeholder image
 
   String get username => _username;
   String get profileImage => _profileImage;
